requirements.txt
pymodbus>=3.0.0
python-dotenv>=0.19.0
pandas>=2.1.0
six>=1.16.0
aiohttp>=3.8.1
matplotlib>=3.7.0
python-dateutil>=2.8.2